+++ elementary-tasks/fibonacci/src/main/java/ssjava/fibonacci/NumberGenerator.java
@@ -4,5 +4,5 @@ package ssjava.fibonacci;
  * Created by Stas on 08.08.2018.
  */
 public interface NumberGenerator {
-	int getNextNumber();
+	long getNextNumber();
 }
